fix: nil exception

+++ app/controllers/api/v2/lodgings_controller.rb
@@ -58,6 +58,6 @@ class Api::V2::LodgingsController < Api::V2::ApiController
       _params = { amenities: true, reviews: true }
 
       return _params.merge(experiences: true, current_user: current_user, lodgings: @lodgings, total_lodgings: @total_lodgings) if action_name == 'index'
-      _params.merge(adults: params[:adults], children: params[:children], nights: (params[:check_out].to_date - params[:check_in].to_date).to_i)
+      _params.merge(adults: params[:adults], children: params[:children], nights: (params[:check_out].to_date - params[:check_in].to_date).to_f)
     end
 end
